Keras-Functional-Model/FaceInsight/render/requirements.txt
tensorflow>=2.15.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pillow>=9.0.0
jinja2>=3.1.0
numpy>=1.23.0
